Instructions
The border color around the text stays blue if the user's current written text is matched correctly with the given text displayed. If the user inputted an incorrect letter or symbol, the border turns red. Finally, when the user successfully finish typing the text with no errors, the border turns green.